Update the API of UsbGetInterface() & UsbGetConfiguration () to follow MdeLib spec.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk@6792 6f19259b-4bc3-4df7-8a09-765794883524

+++ edk2/MdePkg/Include/Library/UefiUsbLib.h
@@ -349,7 +349,7 @@ EFIAPI
 UsbGetInterface (
   IN  EFI_USB_IO_PROTOCOL     *UsbIo,
   IN  UINT16                  Interface,
-  OUT UINT8                   *AlternateSetting,
+  OUT UINT16                  *AlternateSetting,
   OUT UINT32                  *Status
   );
 
@@ -406,7 +406,7 @@ EFI_STATUS
 EFIAPI
 UsbGetConfiguration (
   IN  EFI_USB_IO_PROTOCOL     *UsbIo,
-  OUT UINT8                   *ConfigurationValue,
+  OUT UINT16                  *ConfigurationValue,
   OUT UINT32                  *Status
   );
